support remote SpiceDB
the proxy now take new arguments to configure
a remote SpiceDB:

- spicedb-endpoint: url to the remote SpiceDB
- spicedb-insecure: allows connecting to SpiceDB in plain text
- spicedb-skip-verify-ca: lets the proxy connect to a remote service without
  verifying the certificate trust chain
- spicedb-token: the preshared key of the remote server

if spicedb-endpoint is set to "embedded://", the original
behaviour of spinning up an embedded SpiceDB will take place, which
is useful for testing purposes.

An additional flag to support setting a path for the durable task database is
also added: durabletask-database-path. It will default to /tmp/dtx.sqlite".