changes deployment manifest to use statefulset
With the usage of go-workflows, which we have
currently configured with SQLite, we will need
some durable persistence for the database WAL.

This PR changes the manifest to use a StatefulSet
with 2 replicas for now to demonstrate that
state distributed across replicas should be fine
for this use-case. It also adds a volumeClaimTemplate
for the database storage.