orchestrate certificate generation with cert-manager in dev
until now we've been using Go code to generate certificates
manually in our magefiles. This commit changes to get it
somewhat closer to what we expect it to be like in a production
environment, where we rely on cert-manager.

The cert-managers custom resources here added shall serve
as reference for the production environments

Perhaps the tricky bit here has been retrieving the generated
secrets with the purpose of creating the proxy's upstream kubeconfig.
Alternatively we could pass the arguments to the proxy directly
and have it generate the kube config in memory, which would spare
us some magefile code to do the plumbing.

- The current CRs represent a self-signed issuer, used to then create the client CA
- That is then used to sign a client certificate for a test client, in this case Rakis
- As for the cert used for the HTTP server, we can use the same internal self-signed issuer.
  I discarded using the let's encrypt issuer as it would be a bit of overkill for an internal
  service, and I imagine it wouldn't be possible to do ACME without published DNS records