bug: idempotent delete is not considered successful
if a delete operation over a kube resource is retried,
chances are the resource may have been already deleted
in a previous attempt. In such case Kube will return 404.
However that case wasn't being considered a successful scenario
by the logic, and so the dual-write would try to rollback
the operation. This led to an inconsistent situation where
the kube resource was gone, but the SpiceDB tuples were left
behind.

This manifested during tests with a different response
when calling GetPod over a pod that shouldn't exist:
- when the tuples where not left behind, the proxy
  would respond 403
- when the tuples where left behind, the SpiceDB check
  succeeded, but the upstream returned 404. This
  didn't manifest until now because we didn't
  explicitly handle non 2XX responses in
  AuthzData.FilterResp, see
  https://github.com/authzed/spicedb-kubeapi-proxy/pull/58/commits/a0f9388e0b9ea183fbe0292a9fb79dced0d36575