fixes lack of idempotency in WriteToSpiceDB
I started all of this because of flakes that were trivially
to reproduce, and the underlying issue was WriteToSpiceDB was not idempotent.

It was used by both the optimistic and pesimistic locking workflows, and
the lack of idempotency meant the workflows couldn't recover
by themselves from transient errors. When the tests failed
for the wrong reasons, they flaked.

We introduced `failpoints` as a mechanism to simulate panics
similar to hardware or network failures. However, some recent
changes meant the failpoints were no longer unique and
were causing failures _in the wrong places_.

- when the pessimistic wrote an exclusive lock, retrying it due to
  failure wouldn't work, because the precondition prevented it
  (_make sure no one has a lock on this resource_!)
- when the optimistic wrote with CREATE semantics, a
  subsequent retry after a successful write due to failure
  response failure would fail because the tuples already exist.

This commit proposes making WriteToSpiceDB truly idempotent by
introducing idempotency keys in the SpiceDB schema. All writes
will include a relationships that identify the workflow and the hash
of the payload as the idempotency key.

The flow is as follows:
- perform write
- if failure happens, check if idempotency key was written in previous request
- if exists, assume operation was successful
- if it does not, bubble up the error

The cost of the extra ReadRelationships is only paid in the even of a retry
due to a failure.

The tests were written with the assumption that the system would bubble up
errors after recovery. This goes against the expectations of a durable
workflow engine, which embraces idempotency and is expected to retry
on errors, rather than have the client retry, unless those are unrecoverable.

This was all by design: the workflow wouldn't be responsible to retry things, but rather
execute compensatory operations after an operation failed. This meant
the client had to retry those errors, and it turns out troubleshooting
what happened on a transient error is not that trivial for folks
building on top of the spicedb-kubeapi-proxy.